test(testing): ArticlesScreen tests вынесены в отдельные suite files
state, controller, mutation reducer и presentation tests для ArticlesScreen выделены из RSSReaderTests.swift в самостоятельные feature-oriented suites
