+++ RSSReaderTests/RSSReaderTests.swift
@@ -7,807 +7,6 @@ import Testing
 @MainActor
 struct RSSReaderTests {
     @Test
-    func articlesScreenStateStartsWithoutSelectionPlaceholder() {
-        let state = ArticlesScreenState()
-
-        #expect(state.phase == .noSelection)
-        #expect(state.navigationTitle == "Articles")
-        #expect(state.navigationSubtitle == "0 Unread Items")
-        #expect(state.placeholder?.title == "No Source Selected")
-        #expect(state.toolbarActions.showsSearchAction == false)
-        #expect(state.toolbarActions.showsMarkAllAsReadAction == false)
-    }
-
-    @Test
-    func articlesScreenStateBeginsPrimaryLoadingWhenSelectionChanges() {
-        var state = ArticlesScreenState()
-
-        state.beginLoading(
-            for: .unread,
-            navigationTitle: "Unread",
-            navigationSubtitle: "3 Unread Items",
-            resetsContent: true
-        )
-
-        #expect(state.phase == .loading)
-        #expect(state.navigationTitle == "Unread")
-        #expect(state.navigationSubtitle == "3 Unread Items")
-        #expect(state.showsPrimaryLoadingIndicator)
-        #expect(state.toolbarActions.showsSearchAction == false)
-        #expect(state.toolbarActions.showsMarkAllAsReadAction == false)
-    }
-
-    @Test
-    func articlesScreenStateBuildsEmptyPlaceholderForCurrentSelection() {
-        var state = ArticlesScreenState()
-
-        state.applyLoadedArticles(
-            [],
-            selection: .starred,
-            navigationTitle: "Starred",
-            navigationSubtitle: "0 Starred Items"
-        )
-
-        #expect(state.phase == .empty)
-        #expect(state.navigationTitle == "Starred")
-        #expect(state.navigationSubtitle == "0 Starred Items")
-        #expect(state.placeholder?.title == "No Articles")
-        #expect(state.placeholder?.description == "You have not starred any articles yet.")
-    }
-
-    @Test
-    func articlesScreenStateBuildsPrimaryFailureForInitialLoad() {
-        var state = ArticlesScreenState()
-
-        state.applyLoadingFailure(
-            "Article query service is unavailable.",
-            selection: .inbox,
-            navigationTitle: "All Items",
-            navigationSubtitle: "0 Unread Items",
-            retainsContent: false
-        )
-
-        #expect(state.phase == .failed("Article query service is unavailable."))
-        #expect(state.primaryFailureMessage == "Article query service is unavailable.")
-        #expect(state.refreshFeedback == nil)
-    }
-
-    @Test
-    func articlesScreenStateKeepsVisibleArticlesWhenRefreshFails() {
-        var state = ArticlesScreenState()
-        let unreadItem = makeArticleListItemDTO(isRead: false, isStarred: false)
-
-        state.applyLoadedArticles(
-            [unreadItem],
-            selection: .feed(unreadItem.feedID),
-            navigationTitle: "Feed",
-            navigationSubtitle: "1 Unread Item"
-        )
-        state.beginLoading(
-            for: .feed(unreadItem.feedID),
-            navigationTitle: "Feed",
-            navigationSubtitle: "1 Unread Item",
-            resetsContent: false
-        )
-        state.applyLoadingFailure(
-            "Refresh failed",
-            selection: .feed(unreadItem.feedID),
-            navigationTitle: "Feed",
-            navigationSubtitle: "1 Unread Item",
-            retainsContent: true
-        )
-
-        #expect(state.phase == .loaded)
-        #expect(state.navigationTitle == "Feed")
-        #expect(state.navigationSubtitle == "1 Unread Item")
-        #expect(state.articles.map(\.id) == [unreadItem.id])
-        #expect(state.refreshState == .idle)
-        #expect(state.refreshFeedback == ArticlesScreenRefreshFeedback(message: "Refresh failed"))
-        #expect(state.toolbarActions.isMarkAllAsReadEnabled)
-    }
-
-    @Test
-    func articlesScreenStateClearsRefreshFeedbackWhenPrimaryReloadStarts() {
-        var state = ArticlesScreenState()
-        let unreadItem = makeArticleListItemDTO(isRead: false, isStarred: false)
-
-        state.applyLoadedArticles(
-            [unreadItem],
-            selection: .feed(unreadItem.feedID),
-            navigationTitle: "Feed",
-            navigationSubtitle: "1 Unread Item"
-        )
-        state.presentRefreshFailure("Refresh failed")
-
-        state.beginLoading(
-            for: .unread,
-            navigationTitle: "Unread",
-            navigationSubtitle: "1 Unread Item",
-            resetsContent: true
-        )
-
-        #expect(state.phase == .loading)
-        #expect(state.refreshFeedback == nil)
-    }
-
-    @Test
-    func articlesScreenStateBuildsDerivedToolbarActionsAndSearchPlaceholderForFilteredResults() {
-        var state = ArticlesScreenState()
-        let unreadItem = makeArticleListItemDTO(
-            title: "SwiftUI Weekly",
-            isRead: false,
-            isStarred: false
-        )
-        let readItem = makeArticleListItemDTO(
-            title: "Architecture Digest",
-            isRead: true,
-            isStarred: false
-        )
-
-        state.applyLoadedArticles(
-            [unreadItem, readItem],
-            selection: .feed(unreadItem.feedID),
-            navigationTitle: "Feed",
-            navigationSubtitle: "1 Unread Item"
-        )
-
-        let loadedViewState = state.derivedViewState(searchText: "swift")
-        let emptySearchViewState = state.derivedViewState(searchText: "kotlin")
-
-        #expect(loadedViewState.visibleArticles.map(\.id) == [unreadItem.id])
-        #expect(loadedViewState.toolbarActions.isMarkAllAsReadEnabled)
-        #expect(emptySearchViewState.visibleArticles.isEmpty)
-        #expect(emptySearchViewState.toolbarActions.isMarkAllAsReadEnabled == false)
-        #expect(emptySearchViewState.searchPlaceholder?.title == "No Search Results")
-        #expect(emptySearchViewState.searchPlaceholder?.description == "No visible articles match \"kotlin\".")
-    }
-
-    @Test
-    func articlesScreenStateBuildsRefreshBannerForVisibleRefreshFailure() {
-        var state = ArticlesScreenState()
-        let unreadItem = makeArticleListItemDTO(isRead: false, isStarred: false)
-
-        state.applyLoadedArticles(
-            [unreadItem],
-            selection: .feed(unreadItem.feedID),
-            navigationTitle: "Feed",
-            navigationSubtitle: "1 Unread Item"
-        )
-        state.presentRefreshFailure("Refresh failed")
-
-        let derivedViewState = state.derivedViewState(searchText: "")
-
-        #expect(derivedViewState.refreshBanner?.style == .failed)
-        #expect(derivedViewState.refreshBanner?.title == "Refresh Failed")
-        #expect(derivedViewState.refreshBanner?.message == "Refresh failed")
-        #expect(derivedViewState.refreshBanner?.showsRetryAction == true)
-    }
-
-    @Test
-    func articlesScreenStateBuildsPrimaryLoadingCopyFromSelection() {
-        var state = ArticlesScreenState()
-
-        state.beginLoading(
-            for: .folder("Apple"),
-            navigationTitle: "Apple",
-            navigationSubtitle: "0 Unread Items",
-            resetsContent: true
-        )
-
-        let derivedViewState = state.derivedViewState(searchText: "")
-
-        #expect(derivedViewState.primaryLoadingState?.title == "Loading Articles")
-    }
-
-    @Test
-    func articlesScreenControllerLoadsFeedArticlesForCurrentSelection() async throws {
-        let harness = try TestHarness.make(httpClient: ScriptedHTTPClient())
-        let feed = try #require(try harness.insertFeeds(urls: ["https://example.com/controller-load.xml"]).first)
-        _ = try harness.insertArticle(
-            feed: feed,
-            externalID: "controller-load-article",
-            url: "https://example.com/articles/controller-load",
-            title: "Controller Load"
-        )
-        let controller = ArticlesScreenController()
-
-        await controller.load(
-            selection: .feed(feed.id),
-            sourcesFilter: .allItems,
-            dependencies: harness.dependencies
-        )
-
-        #expect(controller.screenState.phase == .loaded)
-        #expect(controller.screenState.navigationTitle == "controller-load.xml")
-        #expect(controller.screenState.articles.count == 1)
-        #expect(controller.screenState.articles.first?.title == "Controller Load")
-    }
-
-    func articlesScreenControllerPresentsRefreshFailureFromBatchRefreshResult() async throws {
-        let client = ScriptedHTTPClient(
-            responsesByURL: [
-                "https://example.com/controller-refresh.xml": .response(
-                    statusCode: 500,
-                    headers: [:],
-                    body: ""
-                )
-            ]
-        )
-        let harness = try TestHarness.make(httpClient: client)
-        let appState = AppState()
-        let feed = try #require(try harness.insertFeeds(urls: ["https://example.com/controller-refresh.xml"]).first)
-        _ = try harness.insertArticle(
-            feed: feed,
-            externalID: "controller-refresh-article",
-            url: "https://example.com/articles/controller-refresh",
-            title: "Controller Refresh"
-        )
-        let controller = ArticlesScreenController()
-
-        await controller.load(
-            selection: .feed(feed.id),
-            sourcesFilter: .allItems,
-            dependencies: harness.dependencies
-        )
-        harness.dependencies.showFeed(id: feed.id, using: appState)
-
-        await controller.refreshCurrentSelection(
-            selection: .feed(feed.id),
-            dependencies: harness.dependencies,
-            appState: appState
-        )
-
-        #expect(controller.screenState.phase == .loaded)
-        #expect(controller.screenState.refreshFeedback?.message.contains("invalidStatusCode") == true)
-    }
-
-    @Test
-    func articlesScreenControllerClearsPreviousRefreshErrorAfterSuccessfulRefresh() async throws {
-        let client = ScriptedHTTPClient(
-            responsesByURL: [
-                "https://example.com/controller-refresh-success.xml": .response(
-                    statusCode: 200,
-                    headers: [
-                        "Content-Type": "application/rss+xml; charset=utf-8"
-                    ],
-                    body: Self.validRSSFeedXML(
-                        channelTitle: "Refresh Success Feed",
-                        channelLink: "https://example.com/refresh-success/",
-                        language: "en",
-                        itemTitle: "Refreshed Article",
-                        itemLink: "https://example.com/articles/refreshed",
-                        itemGUID: "refreshed-article",
-                        itemDescription: "Readable summary",
-                        pubDate: "Tue, 02 Jan 2024 10:00:00 GMT"
-                    )
-                )
-            ]
-        )
-        let harness = try TestHarness.make(httpClient: client)
-        let appState = AppState()
-        let feed = try #require(try harness.insertFeeds(urls: ["https://example.com/controller-refresh-success.xml"]).first)
-        _ = try harness.insertArticle(
-            feed: feed,
-            externalID: "controller-refresh-success-article",
-            url: "https://example.com/articles/controller-refresh-success",
-            title: "Controller Refresh Success"
-        )
-        let controller = ArticlesScreenController()
-
-        await controller.load(
-            selection: .feed(feed.id),
-            sourcesFilter: .allItems,
-            dependencies: harness.dependencies
-        )
-        controller.screenState.presentRefreshFailure("Previous refresh failed")
-        harness.dependencies.showFeed(id: feed.id, using: appState)
-
-        await controller.refreshCurrentSelection(
-            selection: .feed(feed.id),
-            dependencies: harness.dependencies,
-            appState: appState
-        )
-
-        #expect(controller.screenState.phase == .loaded)
-        #expect(controller.screenState.refreshFeedback == nil)
-    }
-
-    @Test
-    func articlesScreenControllerClearsStaleRefreshErrorWhenSelectionChanges() async throws {
-        let harness = try TestHarness.make(httpClient: ScriptedHTTPClient())
-        let firstFeed = try #require(try harness.insertFeeds(urls: ["https://example.com/controller-selection-a.xml"]).first)
-        let secondFeed = try #require(try harness.insertFeeds(urls: ["https://example.com/controller-selection-b.xml"]).first)
-
-        _ = try harness.insertArticle(
-            feed: firstFeed,
-            externalID: "controller-selection-a-article",
-            url: "https://example.com/articles/controller-selection-a",
-            title: "First Selection"
-        )
-        _ = try harness.insertArticle(
-            feed: secondFeed,
-            externalID: "controller-selection-b-article",
-            url: "https://example.com/articles/controller-selection-b",
-            title: "Second Selection"
-        )
-
-        let controller = ArticlesScreenController()
-
-        await controller.load(
-            selection: .feed(firstFeed.id),
-            sourcesFilter: .allItems,
-            dependencies: harness.dependencies
-        )
-        controller.screenState.presentRefreshFailure("Refresh failed for first feed")
-
-        await controller.load(
-            selection: .feed(secondFeed.id),
-            sourcesFilter: .allItems,
-            dependencies: harness.dependencies
-        )
-
-        #expect(controller.screenState.phase == .loaded)
-        #expect(controller.screenState.selection == .feed(secondFeed.id))
-        #expect(controller.screenState.articles.first?.title == "Second Selection")
-        #expect(controller.screenState.refreshFeedback == nil)
-    }
-
-    @Test
-    func articlesScreenMutationReducerRemovesVisibleArticlesAfterMarkAllAsReadInUnreadFilter() {
-        let firstUnread = makeArticleListItemDTO(isRead: false, isStarred: false)
-        let secondUnread = makeArticleListItemDTO(isRead: false, isStarred: false)
-        let remainingRead = makeArticleListItemDTO(isRead: true, isStarred: false)
-
-        let updatedArticles = ArticlesScreenMutationReducer.reduceAfterMarkAllAsRead(
-            visibleArticles: [firstUnread, secondUnread],
-            allArticles: [firstUnread, secondUnread, remainingRead],
-            filter: ArticleListFilter.unread
-        )
-
-        #expect(updatedArticles.map { $0.id } == [remainingRead.id])
-    }
-
-    @Test
-    func articlesScreenMutationReducerProducesRemoveMutationWhenReadToggleHappensInUnreadFilter() {
-        let unreadArticle = makeArticleListItemDTO(isRead: false, isStarred: false)
-
-        let mutation = ArticlesScreenMutationReducer.mutationAfterToggleReadStatus(
-            article: unreadArticle,
-            filter: ArticleListFilter.unread
-        )
-
-        #expect(mutation == .remove)
-    }
-
-    @Test
-    func articlesScreenMutationReducerProducesUnreadUpdateWhenReadArticleIsToggledBack() {
-        let readArticle = makeArticleListItemDTO(isRead: true, isStarred: false)
-
-        let mutation = ArticlesScreenMutationReducer.mutationAfterToggleReadStatus(
-            article: readArticle,
-            filter: ArticleListFilter.all
-        )
-
-        let updatedArticle: ArticleListItemDTO?
-        if case .update(let article) = mutation {
-            updatedArticle = article
-        } else {
-            updatedArticle = nil
-        }
-
-        #expect(updatedArticle?.isRead == false)
-        #expect(updatedArticle?.isStarred == false)
-    }
-
-    @Test
-    func articlesScreenMutationReducerProducesRemoveMutationWhenUnstarringInsideStarredFilter() {
-        let starredArticle = makeArticleListItemDTO(isRead: true, isStarred: true)
-
-        let mutation = ArticlesScreenMutationReducer.mutationAfterToggleStarred(
-            article: starredArticle,
-            filter: ArticleListFilter.starred
-        )
-
-        #expect(mutation == .remove)
-    }
-
-    @Test
-    func articlesScreenStatePresentsConfirmationOnlyWhenUnreadArticlesAreVisible() {
-        var state = ArticlesScreenState()
-        let readItem = makeArticleListItemDTO(isRead: true, isStarred: false)
-        let unreadItem = makeArticleListItemDTO(isRead: false, isStarred: false)
-
-        state.applyLoadedArticles(
-            [readItem],
-            selection: .feed(readItem.feedID),
-            navigationTitle: "Feed",
-            navigationSubtitle: "0 Unread Items"
-        )
-        state.presentMarkAllAsReadConfirmation()
-        #expect(state.pendingConfirmation == nil)
-
-        state.applyLoadedArticles(
-            [unreadItem],
-            selection: .feed(unreadItem.feedID),
-            navigationTitle: "Feed",
-            navigationSubtitle: "1 Unread Item"
-        )
-        state.presentMarkAllAsReadConfirmation()
-        #expect(state.pendingConfirmation == .markAllAsRead)
-    }
-
-    @Test
-    func articlesScreenControllerPresentsConfirmationWhenAskBeforeMarkingAllAsReadIsEnabled() throws {
-        let harness = try TestHarness.make(httpClient: ScriptedHTTPClient())
-        let appSettingsRepository = try #require(harness.dependencies.appSettingsRepository)
-        let unreadItem = makeArticleListItemDTO(isRead: false, isStarred: false)
-        let controller = ArticlesScreenController(
-            previewScreenState: .previewLoaded(
-                selection: .feed(unreadItem.feedID),
-                navigationTitle: "Feed",
-                navigationSubtitle: "1 Unread Item",
-                articles: [unreadItem]
-            )
-        )
-
-        _ = try appSettingsRepository.update(
-            AppSettingsUpdate(
-                askBeforeMarkingAllAsRead: true,
-                updatedAt: .distantPast
-            )
-        )
-
-        controller.handleMarkAllAsReadAction(
-            searchText: "",
-            selection: .feed(unreadItem.feedID),
-            sourcesFilter: .allItems,
-            dependencies: harness.dependencies,
-            isPreviewMode: false
-        )
-
-        #expect(controller.screenState.pendingConfirmation == .markAllAsRead)
-        #expect(controller.screenState.articles.first?.isRead == false)
-    }
-
-    @Test
-    func articlesScreenControllerMarksAllAsReadImmediatelyWhenAskBeforeMarkingAllAsReadIsDisabled() throws {
-        let harness = try TestHarness.make(httpClient: ScriptedHTTPClient())
-        let appSettingsRepository = try #require(harness.dependencies.appSettingsRepository)
-        let articleStateRepository = try #require(harness.dependencies.articleStateRepository)
-        let unreadItem = makeArticleListItemDTO(isRead: false, isStarred: false)
-        let controller = ArticlesScreenController(
-            previewScreenState: .previewLoaded(
-                selection: .feed(unreadItem.feedID),
-                navigationTitle: "Feed",
-                navigationSubtitle: "1 Unread Item",
-                articles: [unreadItem]
-            )
-        )
-
-        _ = try appSettingsRepository.update(
-            AppSettingsUpdate(
-                askBeforeMarkingAllAsRead: false,
-                updatedAt: .distantPast
-            )
-        )
-
-        controller.handleMarkAllAsReadAction(
-            searchText: "",
-            selection: .feed(unreadItem.feedID),
-            sourcesFilter: .allItems,
-            dependencies: harness.dependencies,
-            isPreviewMode: false
-        )
-
-        let persistedState = try articleStateRepository.fetchStateSnapshot(
-            feedID: unreadItem.feedID,
-            articleExternalID: unreadItem.articleExternalID
-        )
-
-        #expect(controller.screenState.pendingConfirmation == nil)
-        #expect(controller.screenState.articles.first?.isRead == true)
-        #expect(controller.screenState.navigationSubtitle == "0 Unread Items")
-        #expect(persistedState?.isRead == true)
-    }
-
-    @Test
-    func articlesScreenToolbarActionsAreHiddenDuringPrimaryLoading() {
-        var state = ArticlesScreenState()
-
-        state.beginLoading(
-            for: .unread,
-            navigationTitle: "Unread",
-            navigationSubtitle: "0 Unread Items",
-            resetsContent: true
-        )
-
-        #expect(state.toolbarActions.showsSearchAction == false)
-        #expect(state.toolbarActions.showsMarkAllAsReadAction == false)
-    }
-
-    @Test
-    func articlesScreenToolbarActionsAreHiddenAfterPrimaryFailure() {
-        var state = ArticlesScreenState()
-
-        state.applyLoadingFailure(
-            "Unable to load the current selection.",
-            selection: .unread,
-            navigationTitle: "Unread",
-            navigationSubtitle: "0 Unread Items",
-            retainsContent: false
-        )
-
-        #expect(state.toolbarActions.showsSearchAction == false)
-        #expect(state.toolbarActions.showsMarkAllAsReadAction == false)
-    }
-
-    @Test
-    func articlesScreenStateAppliesMarkAllAsReadAndRefreshesToolbarState() {
-        var state = ArticlesScreenState()
-        let unreadItem = makeArticleListItemDTO(isRead: false, isStarred: true)
-
-        state.applyLoadedArticles(
-            [unreadItem],
-            selection: .feed(unreadItem.feedID),
-            navigationTitle: "Feed",
-            navigationSubtitle: "1 Unread Item"
-        )
-        state.presentMarkAllAsReadConfirmation()
-        state.applyMarkAllAsRead(
-            [
-                makeArticleListItemDTO(
-                    id: unreadItem.id,
-                    feedID: unreadItem.feedID,
-                    articleExternalID: unreadItem.articleExternalID,
-                    isRead: true,
-                    isStarred: true
-                )
-            ],
-            navigationSubtitle: "0 Unread Items"
-        )
-
-        #expect(state.pendingConfirmation == nil)
-        #expect(state.phase == .loaded)
-        #expect(state.navigationSubtitle == "0 Unread Items")
-        #expect(state.articles.count == 1)
-        #expect(state.articles.first?.isRead == true)
-        #expect(state.toolbarActions.isMarkAllAsReadEnabled == false)
-    }
-
-    @Test
-    func articlesScreenStateAppliesMarkAllAsReadToUnreadFilterAndTransitionsToEmpty() {
-        var state = ArticlesScreenState()
-        let unreadItem = makeArticleListItemDTO(isRead: false, isStarred: false)
-
-        state.applyLoadedArticles(
-            [unreadItem],
-            selection: .unread,
-            navigationTitle: "Unread",
-            navigationSubtitle: "1 Unread Item"
-        )
-        state.presentMarkAllAsReadConfirmation()
-        state.applyMarkAllAsRead([], navigationSubtitle: "0 Unread Items")
-
-        #expect(state.pendingConfirmation == nil)
-        #expect(state.phase == .empty)
-        #expect(state.navigationSubtitle == "0 Unread Items")
-        #expect(state.articles.isEmpty)
-        #expect(state.toolbarActions.isMarkAllAsReadEnabled == false)
-    }
-
-    @Test
-    func articlesScreenStateAppliesArticleRowUpdateForReadAction() {
-        var state = ArticlesScreenState()
-        let unreadItem = makeArticleListItemDTO(isRead: false, isStarred: false)
-        let updatedItem = makeArticleListItemDTO(
-            id: unreadItem.id,
-            feedID: unreadItem.feedID,
-            articleExternalID: unreadItem.articleExternalID,
-            isRead: true,
-            isStarred: false
-        )
-
-        state.applyLoadedArticles(
-            [unreadItem],
-            selection: .feed(unreadItem.feedID),
-            navigationTitle: "Feed",
-            navigationSubtitle: "1 Unread Item"
-        )
-        state.applyArticleRowMutation(
-            articleID: unreadItem.id,
-            mutation: .update(updatedItem),
-            navigationSubtitle: "0 Unread Items"
-        )
-
-        #expect(state.phase == .loaded)
-        #expect(state.navigationSubtitle == "0 Unread Items")
-        #expect(state.articles.count == 1)
-        #expect(state.articles.first?.isRead == true)
-        #expect(state.toolbarActions.isMarkAllAsReadEnabled == false)
-    }
-
-    @Test
-    func articlesScreenStateRemovesArticleRowForReadActionInUnreadSelection() {
-        var state = ArticlesScreenState()
-        let unreadItem = makeArticleListItemDTO(isRead: false, isStarred: false)
-
-        state.applyLoadedArticles(
-            [unreadItem],
-            selection: .unread,
-            navigationTitle: "Unread",
-            navigationSubtitle: "1 Unread Item"
-        )
-        state.applyArticleRowMutation(
-            articleID: unreadItem.id,
-            mutation: .remove,
-            navigationSubtitle: "0 Unread Items"
-        )
-
-        #expect(state.phase == .empty)
-        #expect(state.navigationSubtitle == "0 Unread Items")
-        #expect(state.articles.isEmpty)
-    }
-
-    @Test
-    func articlesScreenStateAppliesArticleRowUpdateForStarActionOutsideStarredSelection() {
-        var state = ArticlesScreenState()
-        let item = makeArticleListItemDTO(isRead: false, isStarred: false)
-        let updatedItem = makeArticleListItemDTO(
-            id: item.id,
-            feedID: item.feedID,
-            articleExternalID: item.articleExternalID,
-            isRead: false,
-            isStarred: true
-        )
-
-        state.applyLoadedArticles(
-            [item],
-            selection: .feed(item.feedID),
-            navigationTitle: "Feed",
-            navigationSubtitle: "1 Unread Item"
-        )
-        state.applyArticleRowMutation(
-            articleID: item.id,
-            mutation: .update(updatedItem),
-            navigationSubtitle: "1 Unread Item"
-        )
-
-        #expect(state.phase == .loaded)
-        #expect(state.articles.first?.isStarred == true)
-        #expect(state.navigationSubtitle == "1 Unread Item")
-    }
-
-    @Test
-    func articlesScreenStateRemovesArticleRowWhenUnstarredInsideStarredSelection() {
-        var state = ArticlesScreenState()
-        let starredItem = makeArticleListItemDTO(isRead: true, isStarred: true)
-
-        state.applyLoadedArticles(
-            [starredItem],
-            selection: .starred,
-            navigationTitle: "Starred",
-            navigationSubtitle: "1 Starred Item"
-        )
-        state.applyArticleRowMutation(
-            articleID: starredItem.id,
-            mutation: .remove,
-            navigationSubtitle: "0 Starred Items"
-        )
-
-        #expect(state.phase == .empty)
-        #expect(state.navigationSubtitle == "0 Starred Items")
-        #expect(state.articles.isEmpty)
-    }
-
-    @Test
-    func articleRowSwipeActionsStateReflectsReadAndStarredStatus() {
-        let unreadUnstarred = ArticleRowSwipeActionsState(
-            article: makeArticleListItemDTO(isRead: false, isStarred: false)
-        )
-        let readStarred = ArticleRowSwipeActionsState(
-            article: makeArticleListItemDTO(isRead: true, isStarred: true)
-        )
-
-        #expect(unreadUnstarred.readActionTitle == "Read")
-        #expect(unreadUnstarred.readActionSystemImage == "circle")
-        #expect(unreadUnstarred.starActionTitle == "Star")
-        #expect(unreadUnstarred.starActionSystemImage == "star")
-
-        #expect(readStarred.readActionTitle == "Unread")
-        #expect(readStarred.readActionSystemImage == "circle.slash")
-        #expect(readStarred.starActionTitle == "Unstar")
-        #expect(readStarred.starActionSystemImage == "star.slash")
-    }
-
-    @Test
-    func articlesScreenNavigationTitleResolverBuildsTitlesFromSidebarSelection() {
-        #expect(ArticlesScreenNavigationTitleResolver.resolve(selection: nil) == "Articles")
-        #expect(ArticlesScreenNavigationTitleResolver.resolve(selection: .inbox) == "All Items")
-        #expect(ArticlesScreenNavigationTitleResolver.resolve(selection: .unread) == "Unread")
-        #expect(ArticlesScreenNavigationTitleResolver.resolve(selection: .starred) == "Starred")
-        #expect(ArticlesScreenNavigationTitleResolver.resolve(selection: .folder("Tech")) == "Tech")
-        #expect(
-            ArticlesScreenNavigationTitleResolver.resolve(
-                selection: .feed(UUID()),
-                selectedFeedTitle: "The Verge"
-            ) == "The Verge"
-        )
-        #expect(ArticlesScreenNavigationTitleResolver.resolve(selection: .feed(UUID())) == "Source")
-    }
-
-    @Test
-    func articlesScreenSubtitleResolverBuildsSubtitleFromSourcesFilter() {
-        let unreadItem = makeArticleListItemDTO(isRead: false, isStarred: false)
-        let starredItem = makeArticleListItemDTO(isRead: true, isStarred: true)
-        let unreadStarredItem = makeArticleListItemDTO(isRead: false, isStarred: true)
-        let articles = [unreadItem, starredItem, unreadStarredItem]
-
-        #expect(
-            ArticlesScreenSubtitleResolver.resolve(
-                articles: articles,
-                sourcesFilter: .allItems
-            ) == "2 Unread Items"
-        )
-        #expect(
-            ArticlesScreenSubtitleResolver.resolve(
-                articles: articles,
-                sourcesFilter: .unread
-            ) == "2 Unread Items"
-        )
-        #expect(
-            ArticlesScreenSubtitleResolver.resolve(
-                articles: articles,
-                sourcesFilter: .starred
-            ) == "2 Starred Items"
-        )
-    }
-
-    @Test
-    func articlesDaySectionsBuilderGroupsArticlesByDayAndPreservesVisibleOrder() {
-        let calendar = Calendar.current
-        let now = Date()
-        let todayMorning = calendar.date(bySettingHour: 9, minute: 0, second: 0, of: now) ?? now
-        let todayEarlier = calendar.date(bySettingHour: 8, minute: 0, second: 0, of: now) ?? now
-        let yesterdayBase = calendar.date(byAdding: .day, value: -1, to: now) ?? now
-        let yesterdayArticleDate = calendar.date(bySettingHour: 23, minute: 0, second: 0, of: yesterdayBase) ?? yesterdayBase
-
-        let firstToday = makeArticleListItemDTO(title: "Today One", publishedAt: todayMorning)
-        let secondToday = makeArticleListItemDTO(title: "Today Two", publishedAt: todayEarlier)
-        let yesterdayArticle = makeArticleListItemDTO(title: "Yesterday", publishedAt: yesterdayArticleDate)
-
-        let sections = ArticlesDaySectionsBuilder.build(
-            from: [firstToday, secondToday, yesterdayArticle],
-            calendar: calendar
-        )
-
-        #expect(sections.count == 2)
-        #expect(sections[0].articles.map(\.title) == ["Today One", "Today Two"])
-        #expect(sections[1].articles.map(\.title) == ["Yesterday"])
-    }
-
-    @Test
-    func articlesDaySectionsBuilderBuildsTodayYesterdayAndDateHeaders() {
-        let calendar = Calendar.current
-        let now = Date()
-        let today = calendar.startOfDay(for: now)
-        let yesterday = calendar.date(byAdding: .day, value: -1, to: today) ?? today
-        let older = calendar.date(byAdding: .day, value: -2, to: today) ?? today
-
-        #expect(ArticlesDaySectionsBuilder.title(for: today, calendar: calendar) == "Today")
-        #expect(ArticlesDaySectionsBuilder.title(for: yesterday, calendar: calendar) == "Yesterday")
-        #expect(
-            ArticlesDaySectionsBuilder.title(for: older, calendar: calendar)
-            == older.formatted(
-                .dateTime
-                    .weekday(.wide)
-                    .day()
-                    .month(.wide)
-                    .year()
-            )
-        )
-    }
-
-    @Test
     func sourcesFilterPersistencePolicyRestoresPersistedFilterFromSettingsRawValue() {
         let settings = AppSettings(selectedSourcesFilterRawValue: SourcesFilter.starred.rawValue)
